poller/requirements.txt
aio-pika==8.2.2
aiohttp==3.8.3
aiormq==6.4.2
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
pamqp==3.2.1
PyYAML==6.0
yarl==1.8.1
